feat: 위시 다중 삭제 api 연결
