+++ apps/web/src/app/wishlist/_components/WishlistFabArea.tsx
@@ -1,9 +1,11 @@
 import TrashIconFill from '@/assets/icons/fill/trash.svg';
+import Spinner from '@/components/common/spinner';
 
 import WishFab from './WishFab';
 
 type WishlistFabAreaProps = {
   isDeleteMode: boolean;
+  isDeleteWishesPending: boolean;
   isDeleteDisabled: boolean;
   onAddItem: () => void;
   onEnterDeleteMode: () => void;
@@ -12,6 +14,7 @@ type WishlistFabAreaProps = {
 
 function WishlistFabArea({
   isDeleteMode,
+  isDeleteWishesPending,
   isDeleteDisabled,
   onAddItem,
   onEnterDeleteMode,
@@ -25,10 +28,14 @@ function WishlistFabArea({
             type="button"
             onClick={onConfirmDelete}
             aria-label="선택한 위시 삭제하기"
-            disabled={isDeleteDisabled}
+            disabled={isDeleteDisabled || isDeleteWishesPending}
             className="flex size-[62px] cursor-pointer items-center justify-center rounded-full border border-border-neutral-muted bg-bg-layer-default shadow-[0px_0px_8px_0px_rgba(0,0,0,0.04)] disabled:opacity-50"
           >
-            <TrashIconFill width={33} height={33} className="text-icon-neutral-primary" />
+            {isDeleteWishesPending ? (
+              <Spinner size={20} />
+            ) : (
+              <TrashIconFill width={33} height={33} className="text-icon-neutral-primary" />
+            )}
           </button>
         ) : (
           <WishFab onAddItem={onAddItem} onDelete={onEnterDeleteMode} />
